!2171 Fix nullptr in src/daemon/entry * !2160 Fix nullptr in src/daemon/entry

+++ src/daemon/entry/connect/grpc/cri/v1/cri_v1_runtime_runtime_service.cc
@@ -51,6 +51,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::Version(grpc::ServerContext *context,
                                                 runtime::v1::VersionResponse *reply)
 {
     Errors error;
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     m_rService->Version(request->version(), reply, error);
     if (!error.Empty()) {
         return grpc::Status(grpc::StatusCode::UNKNOWN, error.GetMessage());
@@ -64,6 +69,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::CreateContainer(grpc::ServerContext *c
                                                         runtime::v1::CreateContainerResponse *reply)
 {
     Errors error;
+
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
 
     EVENT("Event: {Object: CRI, Type: Creating Container}");
 
@@ -86,6 +96,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::StartContainer(grpc::ServerContext *co
 {
     Errors error;
 
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     EVENT("Event: {Object: CRI, Type: Starting Container: %s}", request->container_id().c_str());
 
     m_rService->StartContainer(request->container_id(), error);
@@ -104,6 +119,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::StopContainer(grpc::ServerContext *con
                                                       runtime::v1::StopContainerResponse *reply)
 {
     Errors error;
+
+    if (request == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
 
     EVENT("Event: {Object: CRI, Type: Stopping Container: %s}", request->container_id().c_str());
 
@@ -124,6 +144,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::RemoveContainer(grpc::ServerContext *c
 {
     Errors error;
 
+    if (request == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     EVENT("Event: {Object: CRI, Type: Removing Container: %s}", request->container_id().c_str());
 
     m_rService->RemoveContainer(request->container_id(), error);
@@ -143,10 +168,15 @@ grpc::Status RuntimeV1RuntimeServiceImpl::ListContainers(grpc::ServerContext *co
 {
     Errors error;
 
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     INFO("Event: {Object: CRI, Type: Listing all Container}");
 
     std::vector<std::unique_ptr<runtime::v1::Container>> containers;
-    m_rService->ListContainers(request->has_filter() ? &request->filter() : nullptr, &containers, error);
+    m_rService->ListContainers(request->has_filter() ? &request->filter() : nullptr, containers, error);
     if (!error.Empty()) {
         ERROR("Object: CRI, Type: Failed to list all containers %s", error.GetMessage().c_str());
         return grpc::Status(grpc::StatusCode::UNKNOWN, error.GetMessage());
@@ -172,6 +202,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::ContainerStats(grpc::ServerContext *co
 {
     Errors error;
 
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     INFO("Event: {Object: CRI, Type: Getting Container Stats: %s}", request->container_id().c_str());
 
     std::unique_ptr<runtime::v1::ContainerStats> contStats =
@@ -193,10 +228,15 @@ grpc::Status RuntimeV1RuntimeServiceImpl::ListContainerStats(grpc::ServerContext
 {
     Errors error;
 
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     INFO("Event: {Object: CRI, Type: Listing all Container stats}");
 
     std::vector<std::unique_ptr<runtime::v1::ContainerStats>> containers;
-    m_rService->ListContainerStats(request->has_filter() ? &request->filter() : nullptr, &containers, error);
+    m_rService->ListContainerStats(request->has_filter() ? &request->filter() : nullptr, containers, error);
     if (!error.Empty()) {
         ERROR("Object: CRI, Type: Failed to list all containers stat %s", error.GetMessage().c_str());
         return grpc::Status(grpc::StatusCode::UNKNOWN, error.GetMessage());
@@ -222,6 +262,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::ContainerStatus(grpc::ServerContext *c
 {
     Errors error;
 
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     INFO("Event: {Object: CRI, Type: Statusing Container: %s}", request->container_id().c_str());
 
     std::unique_ptr<runtime::v1::ContainerStatus> contStatus =
@@ -243,6 +288,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::ExecSync(grpc::ServerContext *context,
 {
     Errors error;
 
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     WARN("Event: {Object: CRI, Type: sync execing Container: %s}", request->container_id().c_str());
 
     m_rService->ExecSync(request->container_id(), request->cmd(), request->timeout(), reply, error);
@@ -261,6 +311,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::RunPodSandbox(grpc::ServerContext *con
                                                       runtime::v1::RunPodSandboxResponse *reply)
 {
     Errors error;
+
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
 
     EVENT("Event: {Object: CRI, Type: Running Pod}");
 
@@ -282,6 +337,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::StopPodSandbox(grpc::ServerContext *co
 {
     Errors error;
 
+    if (request == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     EVENT("Event: {Object: CRI, Type: Stopping Pod: %s}", request->pod_sandbox_id().c_str());
 
     m_rService->StopPodSandbox(request->pod_sandbox_id(), error);
@@ -302,6 +362,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::RemovePodSandbox(grpc::ServerContext *
 {
     Errors error;
 
+    if (request == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     EVENT("Event: {Object: CRI, Type: Removing Pod: %s}", request->pod_sandbox_id().c_str());
 
     m_rService->RemovePodSandbox(request->pod_sandbox_id(), error);
@@ -321,6 +386,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::PodSandboxStatus(grpc::ServerContext *
                                                          runtime::v1::PodSandboxStatusResponse *reply)
 {
     Errors error;
+
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
 
     INFO("Event: {Object: CRI, Type: Status Pod: %s}", request->pod_sandbox_id().c_str());
 
@@ -343,6 +413,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::ListPodSandbox(grpc::ServerContext *co
                                                        runtime::v1::ListPodSandboxResponse *reply)
 {
     Errors error;
+
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
 
     INFO("Event: {Object: CRI, Type: Listing all Pods}");
 
@@ -373,6 +448,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::PodSandboxStats(grpc::ServerContext *c
 {
     Errors error;
 
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     INFO("Event: {Object: CRI, Type: Stats Pod: %s}", request->pod_sandbox_id().c_str());
 
     std::unique_ptr<runtime::v1::PodSandboxStats> podStats;
@@ -396,10 +476,15 @@ RuntimeV1RuntimeServiceImpl::ListPodSandboxStats(grpc::ServerContext *context,
 {
     Errors error;
 
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     INFO("Event: {Object: CRI, Type: Listing Pods Stats}");
 
     std::vector<std::unique_ptr<runtime::v1::PodSandboxStats>> podsStats;
-    m_rService->ListPodSandboxStats(request->has_filter() ? &request->filter() : nullptr, &podsStats, error);
+    m_rService->ListPodSandboxStats(request->has_filter() ? &request->filter() : nullptr, podsStats, error);
     if (!error.Empty()) {
         ERROR("Object: CRI, Type: Failed to list pods stats: %s", error.GetCMessage());
         return grpc::Status(grpc::StatusCode::UNKNOWN, error.GetMessage());
@@ -425,6 +510,11 @@ RuntimeV1RuntimeServiceImpl::UpdateContainerResources(grpc::ServerContext *conte
 {
     Errors error;
 
+    if (request == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     WARN("Event: {Object: CRI, Type: Updating container resources: %s}", request->container_id().c_str());
 
     m_rService->UpdateContainerResources(request->container_id(), request->linux(), error);
@@ -445,6 +535,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::Exec(grpc::ServerContext *context,
 {
     Errors error;
 
+    if (request == nullptr || response == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     EVENT("Event: {Object: CRI, Type: execing Container: %s}", request->container_id().c_str());
 
     m_rService->Exec(*request, response, error);
@@ -464,6 +559,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::Attach(grpc::ServerContext *context,
                                                runtime::v1::AttachResponse *response)
 {
     Errors error;
+
+    if (request == nullptr || response == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
 
     EVENT("Event: {Object: CRI, Type: attaching Container: %s}", request->container_id().c_str());
 
@@ -486,6 +586,11 @@ RuntimeV1RuntimeServiceImpl::UpdateRuntimeConfig(grpc::ServerContext *context,
 {
     Errors error;
 
+    if (request == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
+
     EVENT("Event: {Object: CRI, Type: Updating Runtime Config}");
 
     m_rService->UpdateRuntimeConfig(request->runtime_config(), error);
@@ -504,6 +609,11 @@ grpc::Status RuntimeV1RuntimeServiceImpl::Status(grpc::ServerContext *context,
                                                runtime::v1::StatusResponse *reply)
 {
     Errors error;
+
+    if (request == nullptr || reply == nullptr) {
+        ERROR("Invalid input arguments");
+        return grpc::Status(grpc::StatusCode::INVALID_ARGUMENT, "Invalid input arguments");
+    }
 
     INFO("Event: {Object: CRI, Type: Statusing daemon}");
